Add old progress data type support for reporting configurations (#226)
* Add old progress data type support for reporting configurations

Some Enterprise reporting configs have an old reporting data type that
we don't want to create more of. We still need to maintain these
reports, so they will still show up in the list of configs. You will not
be able to change the type of the report, so if a different type of
report is needed, you must create a new one.

* Added test make sure data type is not disabled for newer formats.