fix: course about URL should use course key, not course run key.
When the spent table was first implemented, it utilized the Analytics
API to fetch enrollments, and that data source contained a course key
field.  However, when we switched over to real-time fetching of
redemptions via the transaction list API for assignable budgets, we
erroneously treated the `content_key` field as a course key, breaking
the course about page link in the spent table.

The following two PRs implment a `parent_content_key` field in the
transaction list API, which we can now utilize for the real-time spent
table implmentation:

* https://github.com/openedx/openedx-ledger/pull/60
* https://github.com/openedx/enterprise-subsidy/pull/209

ENT-8389