fix: consider mixed state of budget expiries when rendering expiry modal
When determining budget expiry thresholds/modals, We need to consider the special case where
there are a mix of expired and non-expired budgets.  In that case, we only want
to determine the expiry threshold from the set of *non-expired* budgets,
so that the expiry alert and modal below do not falsely indicate that *all* budgets are expired.
ENT-8834